fix(pages): refactor stored and saved codes page
