+++ src/lib/components/table/transactions/TransactionsTable.tsx
@@ -25,9 +25,9 @@ export const TransactionsTable = ({
   if (isLoading) return <Loading withBorder />;
   if (!transactions?.length) return emptyState;
 
-  const templateColumns = `25px 180px 48px minmax(360px, 1fr) ${
+  const templateColumns = `40px 190px 48px minmax(360px, 1fr) ${
     showRelations ? "100px " : ""
-  }max(160px) ${showTimestamp ? "max(220px) " : ""}${
+  }max(160px) ${showTimestamp ? "max(230px) " : ""}${
     showAction ? "100px " : ""
   }`;
