+++ src/lib/pages/stored-codes/components/SaveCodeButton.tsx
@@ -1,0 +1,12 @@
+import { CustomIcon } from "lib/components/icon";
+import { SaveNewCodeModal } from "lib/components/modal/code/SaveNewCode";
+
+export const SaveCodeButton = () => (
+  <SaveNewCodeModal
+    buttonProps={{
+      variant: "outline-primary",
+      leftIcon: <CustomIcon name="bookmark" />,
+      children: "Save New Code",
+    }}
+  />
+);
