--- src/pages/[network]/my-codes.tsx
@@ -1,3 +1,0 @@
-import MyCodes from "lib/pages/my-codes";
-
-export default MyCodes;
