+++ src/pages/saved-codes.tsx
@@ -1,0 +1,3 @@
+import SavedCodes from "lib/pages/saved-codes";
+
+export default SavedCodes;
